requirements.txt
asgiref==3.4.1
beautifulsoup4==4.11.1
certifi==2021.10.8
charset-normalizer==2.0.12
Django==3.2.13
django-author==1.2.0
django-bootstrap4==21.1
django-currentuser==0.5.3
django-paypal==2.0
idna==3.3
importlib-metadata==2.1.3
paypal==1.2.5
Pillow==8.4.0
pkg_resources==0.0.0
psycopg2==2.9.3
pytz==2022.1
requests==2.27.1
setuptools-git==1.2
soupsieve==2.3.2
sqlparse==0.4.2
typing_extensions==4.1.1
urllib3==1.26.9
zipp==3.6.0
